refactor getCaptures
