+++ app/test/unit/helpers/regex-test.ts
@@ -1,4 +1,4 @@
-import { getCaptures } from '../../../src/lib/helpers/get-captures'
+import { getCaptures } from '../../../src/lib/helpers/regex'
 
 describe('getCaptures()', () => {
   let bodyOfText: string
@@ -9,8 +9,8 @@ describe('getCaptures()', () => {
     beforeAll(() => {
       bodyOfText = `capture me!:matching:capture me too!\nalso capture me!:matching:also capture me too!\n`
     })
-    it('returns all captures', () => {
-      expect(subject()).toEqual(
+    it('returns all captures', async () => {
+      expect(await subject()).toEqual(
         new Set<Array<string>>([
           ['capture me!', 'capture me too!'],
           ['also capture me!', 'also capture me too!'],
@@ -22,8 +22,8 @@ describe('getCaptures()', () => {
     beforeAll(() => {
       bodyOfText = ' '
     })
-    it('returns empty set', () => {
-      expect(subject()).toEqual(new Set<Array<string>>())
+    it('returns empty set', async () => {
+      expect(await subject()).toEqual(new Set<Array<string>>())
     })
   })
 })
